third_party/blink/web_tests/fast/forms/range/range-drag-vertical-expected.txt

This is a testharness.js-based test.
[FAIL] writing-mode:sideways-lr, direction:rtl, appearance:none; All input[range] with vertical writing mode should change value on drag.
  assert_equals: drag slider thumb up will decrease value. expected "0" but got "100"
[FAIL] writing-mode:sideways-lr, direction:rtl, appearance:auto; All input[range] with vertical writing mode should change value on drag.
  assert_equals: drag slider thumb up will decrease value. expected "0" but got "100"
[FAIL] writing-mode:sideways-lr, direction:ltr, appearance:none; All input[range] with vertical writing mode should change value on drag.
  assert_equals: drag slider thumb up will increase value. expected "100" but got "0"
[FAIL] writing-mode:sideways-lr, direction:ltr, appearance:auto; All input[range] with vertical writing mode should change value on drag.
  assert_equals: drag slider thumb up will increase value. expected "100" but got "0"
Harness: the test ran to completion.
